requirements.txt
# Web框架
fastapi==0.109.2
uvicorn==0.27.1

# 数据库
sqlalchemy==2.0.27
python-dotenv==1.0.1

# 爬虫相关
selenium==4.24.0
opencv-python==4.9.0.80
numpy==1.24.4
pillow==10.4.0

# 定时任务
apscheduler==3.10.4

# 工具库
tenacity==8.2.3
pydantic==2.6.1

# 开发工具
python-multipart==0.0.9